requirements.txt
streamlit>=1.39.0
pandas>=2.0.0
numpy>=1.26.0
python-dateutil>=2.8.2
pydantic>=2.0.0
pytest>=7.0.0
